backend/adk_service/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
httpx>=0.28.0
pydantic>=2.11.0
google-adk>=1.0.0
google-genai>=1.30.0
